Fix decorator support
