+++ .babelrc.js
@@ -1,6 +1,10 @@
 module.exports = {
   presets: ['next/babel'],
   plugins: [
+    'babel-plugin-transform-typescript-metadata',
+    ['@babel/plugin-proposal-decorators', { legacy: true }],
+    ['@babel/plugin-proposal-class-properties', { loose: true }],
+    'babel-plugin-parameter-decorator',
     // Resolve imports to top level directory for better tree shaking.
     [
       'babel-plugin-transform-imports',
