+++ package.json
@@ -23,7 +23,7 @@
     "dev:server": "ttsc --project ./server --watch",
     "build": "yarn build:server && yarn build:app",
     "build:app": "next build",
-    "build:server": "ttsc --project ./server",
+    "build:server": "rm -rf ./server-dist && ttsc --project ./server",
     "analyze": "cross-env ANALYZE=true yarn build:app",
     "generate": "prisma generate",
     "migrate:save": "prisma migrate dev --preview-feature",
@@ -46,7 +46,6 @@
     "clsx": "^1.1.1",
     "cookie": "^0.4.1",
     "deepmerge": "^4.2.2",
-    "email-validator": "^2.0.4",
     "formik": "^2.2.6",
     "graphql": "^15.4.0",
     "immutability-helper": "^3.1.1",
@@ -71,6 +70,8 @@
   },
   "devDependencies": {
     "@babel/core": "^7.12.10",
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/plugin-proposal-decorators": "^7.12.12",
     "@next/bundle-analyzer": "^10.0.4",
     "@prisma/cli": "^2.13.1",
     "@types/bcryptjs": "^2.4.2",
@@ -80,7 +81,6 @@
     "@types/js-cookie": "^2.2.6",
     "@types/jsonwebtoken": "^8.5.0",
     "@types/lodash": "^4.14.165",
-    "@types/module-alias": "^2.0.0",
     "@types/node": "^14.14.16",
     "@types/nodemailer": "^6.4.0",
     "@types/nprogress": "^0.2.0",
@@ -93,7 +93,9 @@
     "@typescript-eslint/eslint-plugin": "^4.11.0",
     "@typescript-eslint/parser": "^4.11.0",
     "babel-eslint": "^10.1.0",
+    "babel-plugin-parameter-decorator": "^1.0.16",
     "babel-plugin-transform-imports": "^2.0.0",
+    "babel-plugin-transform-typescript-metadata": "^0.3.1",
     "cross-env": "^7.0.3",
     "dedent": "^0.7.0",
     "eslint": "^7.16.0",
