+++ pages/api/graphql.ts
@@ -1,5 +1,13 @@
+// Required for TypeGraphQL
+import 'reflect-metadata';
+
+import { ApolloServer } from 'apollo-server-micro';
 import { NextApiHandler } from 'next';
-import { createServer } from '~server/apollo-server';
+import { buildSchema } from 'type-graphql';
+
+import { authChecker } from '~server/auth-checker';
+import { createContext } from '~server/context';
+import { UserResolver } from '~server/schemas';
 
 // Next.js API route config
 // https://nextjs.org/docs/api-routes/api-middlewares
@@ -9,6 +17,10 @@ export const config = {
   },
 };
 
+// Check environment.
+const isDevelopment = process.env.NODE_ENV === 'development';
+
+// Cache route handler.
 let handler: NextApiHandler;
 
 const apiHandler: NextApiHandler = async (req, res) => {
@@ -16,8 +28,27 @@ const apiHandler: NextApiHandler = async (req, res) => {
     return handler(req, res);
   }
 
-  const apolloServer = await createServer();
-  handler = apolloServer.createHandler({ path: '/api/graphql' });
+  // Build TypeGraphQL executable schema.
+  const schema = await buildSchema({
+    resolvers: [UserResolver],
+    authChecker,
+  });
+
+  // Create Apollo server.
+  const apolloServer = new ApolloServer({
+    schema,
+    context: createContext,
+    // Enable playground in development mode.
+    playground: isDevelopment,
+    introspection: isDevelopment,
+    // Include data tracing.
+    tracing: isDevelopment,
+    // mocks: isDevelopment,
+  });
+
+  handler = apolloServer.createHandler({
+    path: '/api/graphql',
+  });
 
   return handler(req, res);
 };
