+++ server/tsconfig.json
@@ -1,24 +1,24 @@
 {
   "compilerOptions": {
+    "baseUrl": ".",
     "declaration": true,
+    "emitDecoratorMetadata": true,
+    "esModuleInterop": true,
+    "experimentalDecorators": true,
+    "forceConsistentCasingInFileNames": true,
     "incremental": true,
     "module": "commonjs",
-    "target": "es2020",
-    "strict": true,
-    "strictPropertyInitialization": false,
-    "baseUrl": ".",
-    "esModuleInterop": true,
     "moduleResolution": "Node",
     "outDir": "../server-dist",
     "paths": {
       "~server/*": ["./*"],
     },
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
+    "plugins": [{ "transform": "typescript-transform-paths" }],
     "resolveJsonModule": true,
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "plugins": [{ "transform": "typescript-transform-paths" }]
+    "skipLibCheck": true,
+    "strict": true,
+    "strictPropertyInitialization": false,
+    "target": "es2020"
   },
   "include": ["../global.d.ts", "**/*.ts"]
 }
