+++ tsconfig.json
@@ -1,24 +1,27 @@
 {
   "compilerOptions": {
     "allowJs": true,
+    "allowSyntheticDefaultImports": true,
+    "baseUrl": ".",
+    "emitDecoratorMetadata": true,
+    "esModuleInterop": true,
+    "experimentalDecorators": true,
+    "forceConsistentCasingInFileNames": true,
+    "isolatedModules": true,
     "jsx": "preserve",
     "lib": ["dom", "dom.iterable", "esnext"],
     "module": "commonjs",
-    "target": "es6",
-    "noEmit": true,
-    "isolatedModules": true,
-    "strict": true,
-    "allowSyntheticDefaultImports": true,
-    "baseUrl": ".",
-    "esModuleInterop": true,
     "moduleResolution": "Node",
+    "noEmit": true,
     "paths": {
       "~app/*": ["app/*"],
       "~server/*": ["server/*"],
     },
+    "resolveJsonModule": true,
     "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true
+    "strict": true,
+    "strictPropertyInitialization": false,
+    "target": "es6"
   },
   "exclude": ["node_modules"],
   "include": [
